iterating based on feedback
